endpints secured

+++ src/main/java/com/booking/bookingApp/service/ReviewService.java
@@ -1,21 +1,30 @@
 package com.booking.bookingApp.service;
+import com.booking.bookingApp.config.JwtService;
 import com.booking.bookingApp.dto.ReviewDto;
 import com.booking.bookingApp.entity.Favourite;
 import com.booking.bookingApp.entity.Review;
+import com.booking.bookingApp.entity.User;
 import com.booking.bookingApp.repository.ReviewRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
+import java.util.Collection;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @RequiredArgsConstructor
 @Service
 public class ReviewService {
     private final ReviewRepository reviewRepository;
+    private final JwtService jwtService;
 
     public Review saveReview(Review review) {
         review.setCreatedAt(LocalDate.now());
@@ -23,8 +32,13 @@ public class ReviewService {
     }
 
     public Review updateReview(Review review) {
-        if (reviewRepository.findById(review.getId()).isPresent()) {
-            return reviewRepository.save(review);
+        Optional<Review> response = reviewRepository.findById(review.getId());
+        if (response.isPresent()) {
+            String email = response.get().getUser().getEmail();
+            String username = jwtService.getUsername();
+            if (username.equals(email)) {
+                return reviewRepository.save(review);
+            } else return null;
         } else return null;
     }
 
@@ -48,17 +62,20 @@ public class ReviewService {
         return reviewDtos;
     }
 
-    public void deleteReview(Long id) {
-        reviewRepository.deleteById(id);
+    public void deleteReview(Long id) throws Exception {
+        Optional<Review> review = reviewRepository.findById(id);
+        if (review.isPresent()) {
+            String email = review.get().getUser().getEmail();
+            String username = jwtService.getUsername();
+            List<String> roles = jwtService.getRolesFromJwt();
+            if (roles.contains("ROLE_ADMIN") || email.equals(username)) {
+                reviewRepository.deleteById(id);
+            } else throw new Exception("La review no fue elieminada.");
+        }
     }
 
-//    @Transactional
-//    public void deleteList(Iterable<Review> reviews) {
-//        reviewRepository.deleteAll(reviews);
-//    }
-
     private ReviewDto reviewToReviewDto(Review review) {
-        ReviewDto.ReviewUser user = new ReviewDto.ReviewUser(review.getUser().getId(), review.getUser().getFirstName(), review.getUser().getLastName());
+        ReviewDto.ReviewUser user = new ReviewDto.ReviewUser(review.getUser().getId(), review.getUser().getFirstName(), review.getUser().getLastName(), review.getUser().getImageUrl());
         return new ReviewDto(
                 review.getId(),
                 review.getDescription(),
